Streamline configuration of account/admin hosts
There is one reasonable default for the tenant domain template, if only
the admin_host is configured, namely `"%{tenant}.#{admin_host}"`.
Adding one piece of logic in code means we no longer have to specify a
base in `settings.yml` only to override it.  This flexibility will be
useful when we try to standardize development with multitenancy enabled.

This is backwards compatible and all previous configs will still work.