Overhauled Docker configuration and modified specs to allow to run inside Docker
This commit is in preperation of beginning to release docker images that can be used via Docker Hub with each new release. It also allows for fully dockerized development environment and production environments.

Two phase Docker build - base and main. The goal here are much faster build times by moving seldom changing items to the base image.

Moves to Passenger base image which allows a few real niceties:

  - for dev, you can restart the web process w/o restarting the container, saving a lot of time
  - for prod passenger is more effecient by default in a variety of configurations than puma. Puma frequenty requires more tuning
  - means nginx is serving assets, so no extra proxy is required

Note, this makes the build a little bigger, but with libreoffice and fits in there this change is a pretty small fraction in my opinion.

Uses a nulldb to compile assets during build, making deployment much faster. Also set up to allow for optional assets volume, which makes keeping cached assets possible

This Dockerfile moves fits to always be in the same place in the container. There is still a version file present, so we don't really need it in the path and that means we don't have to update multiple places when we change versions. Just
 in the Dockerfile.base

Sets the admin and default host in dockerfile and virtual_host attributes. That allows this setup to work in dev by default with common proxies like Dory and others.

Makes test adjustments to allow for Docker environment, adds an external Chrome driver docker container to the dev environment

Splits the production and dev docker-compose.yml files. Reduced dev zookeeper to 1, removed lb and memcached from dev. Splitting futher refinement of production docker-compose in to a seperate task.