Move SOLR from Bitnami to address M1 chip issue
These changes are cribbed from [Adventist's b53bbc6b commit][1].

Prior to this commit, when I ran `sc up` the script exited with the
following message:

> container for service "solr" is unhealthy

With this commit, when I run `sc up` the script successfully boots up
the stack and I can go to `http://admin.bl.test`.

One caveat is that there is also a `docker-compose.production.yml`
file.  I have not cross-checked that file with the `docker-compose.yml`
changes; this is not a space I am yet prepared to explore (nor do I have
adequate knowledge to say whether or not this needs changing and what to
change).

Relates to:

- #201

[1]: https://github.com/scientist-softserv/adventist-dl/commit/b53bbc6b3b827423f8b7486f60b544ad7b593af5