Conditionally render BrowseEverything based on config
Prior to this commit, the only time we would render the BrowseEverything
link was when someone requested the page with the `params[:cloud]`
parameter.

However, as we look to configure BrowseEverything to use remotes, that
conditional would be a stumbling block.

With this commit, we now inspect the Hyrax and BrowseEverything
configuration to see if we should render the widget.  As a courtesy, I
have preserved the `params[:cloud]` as a short-cut.  In fact, that
courtesy makes it far easier to test if the buttons are present.

/Note:/ this does not resolve of whether or not it works to select a
file via browse everything and then "process" that file to attach to a
work.

Related to:

- #190