Adding override to bug in Bulkrax v4.4.0
Prior to this commit, the upstream logic was, by way of inheritance, as
follows:

```ruby
def create_objects(types = [])
  types.each do |object_type|
    parser.send("create_#{object_type.pluralize}")
  end
end
```

Note that in the above code, we have the `parser.send` call.  The
`parser` is an undefined method/local variable.

With this commit, we're making an explicit override that is introduced
in [samvera-labs/bulkrax#713][713].

Considering that we're approaching a Bulkrax upgrade, this is an
interstitial change.

[713]: https://github.com/samvera-labs/bulkrax/pull/713