Handling more fields with XML Parser
Prior to this commit, nothing was calling `add_complicated_fields`.  As
a result some of the intended logic was never executed.

With this commit, we're handling a few more parser issues.

It does not, however, address other issues (such as calling the
`add_fields_with_names`).  That will need to be resolved elsewhere.

What it is, however, highlighting is that the `build_metadata` method
warrants significant reconsideration; namely it is hard to test.

_Musings_

What might that test look like?  We know that after calling
`build_metadata` we want to set the `parse_metadata` value.  What goes
into that parsing?  The raw metadata and the parser mappings (and
perhaps other things).  It would be somewhat easy to extract a class in
Bulkrax that we could pass in the raw metadata (or whatever assumptive
structure) and the parser mappings.

There remains the challenge of how we determine the factory_class, but
that's also a smaller problem.

In other words, given the raw metadata and parser, we expected the
following parsed_metadata.  Without this test, we are doomed to play a
game of whack-a-mole.

Related to:

- https://github.com/scientist-softserv/britishlibrary/issues/174
- https://github.com/scientist-softserv/britishlibrary/issues/178
- https://github.com/scientist-softserv/britishlibrary/issues/186