Adding XML parsing for creator and contributor
This commit has five concerns:

1. Removing an unnecessary call to `String#constantize`
2. Moving methods for greater visibility
3. Adding some todo items for future consideration
4. Replacing `map.each` chain with `each`
5. Parsing creator and contributor

I consider removing unnecessary call to `String#constantize` to be
self-explanatory.

For moving methods, I wanted the "add_complicated_fields" and the
"complicated_elements" to be in close proximity.  This helps connect the
future reader to information.

I added the todo items because I think it's a bit disjointed having this
XML specific information tucked into a class when other aspects are
found in the Bulkrax field mappings.

Replacing `map.each` chain with `each` improves removes one loop (or
conversion of an array into an iterator).

Last, parsing the creator and contributor.  Or, what all we're here
for.  Without these changes, we only parse a portion of the
creator (e.g. their ISNI and/or ORCID).

With these changes we begin parsing the creator as identified in the
specs provided.  I believe the JSON object as string is the correct
format based on examing other complicated fields that were previously
parsed and rendered (e.g. `current_he_institution`).

I will need to verify this encoding with a full import.  Why a full
import?  Because I need to see the persisted Fedora object loaded and
examine it's metadata properties.

Related to:

- https://github.com/scientist-softserv/britishlibrary/issues/174
- https://github.com/scientist-softserv/britishlibrary/issues/177
- https://github.com/scientist-softserv/britishlibrary/issues/178
- https://github.com/scientist-softserv/britishlibrary/issues/181
- https://github.com/scientist-softserv/britishlibrary/issues/186