Handling multiple values for "people" fields
Prior to this commit, we were only handling the last person encountered
in the parsing.

In adding the third parameter (e.g. position) to the `#add_metadata`
call we're building a more complex object; and preventing the last
element of a multi-value field from obliterating the previous elements.

With this commit, we're also adding a new XML element to test (one that
has two people) and verifying that the parsed_metadata is as expected.

Note, there were several of us looking on as we then went through my
local environment to look at the importer's parsed and then rendered
metadata.

Related to:

- https://github.com/scientist-softserv/britishlibrary/issues/174
- https://github.com/scientist-softserv/britishlibrary/issues/181