Capturing contributor_type for XML import
This commit contains two things:

1. Method definitions for expected implementations in Bulkrax.
2. Adding mapping to contributor_type for ETD's XML metadata.

**Method Definitions**

I encountered `UndefinedMethodError` when I was re-running an importer.
In [Bulkrax's XML Parser (as of c3a63813)][1] there are method
definitions that override the [Application Parser's methods][2].  In
updating this, we conform to the interface and implement the methods of
that interface.

**Mapping to contributor_type**

Prior to this commit, our [CSV parser handled the contributor_type][3]
but our XML parser did not.

With this commit, our XML parser now considered `contributor_type` a
valid field.  And our XML entry now appends that field value to the
complex contributor object.  I have also put in place a guard condition,
in part to assure that we have a valid entry; but also to highlight a
connect between this field and the UI where we set the field.

Last we need to verify that this change propagates to the UI, both show
and edit page; I did this locally but our QA will need to handle that.

I referenced http://ethos.bl.uk/ProcessSearch.do?query=629358 for this
test.

Related to:

- https://github.com/scientist-softserv/britishlibrary/issues/181

[1]: https://github.com/samvera-labs/bulkrax/blob/c3a638139be3ec537aaee250489c7e3c8a4309f3/app/parsers/bulkrax/xml_parser.rb#L9-L19)

[2]: https://github.com/samvera-labs/bulkrax/blob/c3a638139be3ec537aaee250489c7e3c8a4309f3/app/parsers/bulkrax/application_parser.rb#L167-L184

[3]: https://github.com/scientist-softserv/britishlibrary/blob/b73202952cc384743df4079b868366dbedc65fe1/config/initializers/bulkrax.rb#L79