[i227] Addressing Tenant security issue in Bulkrax
This commit does one primary thing, leverages Bulkrax to v4.4.0+.

We are in the process of releasing Hyrax 5.0; however with a reported
issue in the British Library repository, we want to explore using a
version that should resolve that bug; without waiting for a release.

From v4.4.0 to the ref `b8905d8644617d23b97a7346dba4a223537502d1` there
appears to be only one breaking change; namely the change for adding
permissions to Bulkrax.

Related to:

- https://github.com/samvera-labs/bulkrax/pull/707
- https://github.com/scientist-softserv/britishlibrary/issues/227