Establishing imported model before metadata
Prior to this commit, given the Bulkrax configuration, if we imported a
CSV with a `work_type` column name (and not `model` column) we would get
the following error:

```
NoMethodError:
  undefined method `method_defined?' for nil:NilClass
  # ./app/models/concerns/bulkrax/has_matchers.rb:128:in `field_supported?'
  # ./app/models/concerns/bulkrax/has_matchers.rb:157:in `multiple?'
  # ./app/models/concerns/bulkrax/has_matchers.rb:34:in `block in add_metadata'
  # ./app/models/concerns/bulkrax/has_matchers.rb:31:in `each'
  # ./app/models/concerns/bulkrax/has_matchers.rb:31:in `add_metadata'
  # ./app/models/bulkrax/csv_entry.rb:77:in `block in add_ingested_metadata'
  # ./app/models/bulkrax/csv_entry.rb:75:in `each'
  # ./app/models/bulkrax/csv_entry.rb:75:in `add_ingested_metadata'
  # ./app/models/bulkrax/csv_entry.rb:43:in `build_metadata'
```

This was because Bulkrax assumes `model` as the column name but provides
a means of specifying otherwise.  The receiver of `method_defined?` is
`nil` but the variable name is `factory_class`.

After this commit, before we attempt to parse the metadata we set the
model and thus the `factory_class`.

Related to:

- https://github.com/samvera-labs/bulkrax/pull/719
- https://github.com/samvera-labs/bulkrax/issues/718
- https://github.com/scientist-softserv/britishlibrary/issues/211