Fixed add to collection button
Before this commit, the JS assumed the a section tag with data
attributes. Without that, the button did not work.

With this commit, the data attributes were added as well as a partial to
restore functionality of the add to collection button.

We also needed to ensure we added the modal_add_to_collection from:
https://github.com/scientist-softserv/utk-hyku/blame/d6260ee598f11ae4ee201d0880ff0d893447e271/app/views/hyrax/dashboard/collections/_show_actions.html.erb#L11

Ref: #133