Refactor `DisplaysContentDecorator`
I feel I was a bit to heavy handed and presumptive about the changes to
`DisplaysContent`.  I'm putting back in the webm and ogg since this is
how the Hyrax IIIF AV gem serves it up and will let the future decide if
those formats are necessary or not.  Also, instead of overriding methods
to change values (ex. in the gem it calls `solr_document` but in Hyku,
that same value is called `object`), I'm making almost alias like
methods to return what is expected.  However,I'm unsure if this is the
proper way of doing it.