Modify sanitization for version 3 manifests labels
For Presentation 3 manifests, the sanitization of values seemed to bug
out and return something like `{\"@none\"=&gt;[\"This is a title\"]}`
for the label of the manifest.  It looks like a literal hash that was
turned into a string.  We just want `This is a title` in the manifest
instead.  This commit will unescape some characters to make that happen.