Hot Fix: "undefined local variable or method `section`"
[In Slack][1] this was identified as something to quickly fix.

The underlying problem is that the `section` local for the "_creator"
partial was not being passed to this partial.

The failure comes from the [Hyrax::PresenterRender][2] not passing the
`section` as a local variable to the template.  To unblock the client,
I'm adding a check for `section` before attempting to test equality.

Below are the three `app/views` that have a `section ==` logic check.

For this commit, I'm amending `_contributor` and `_creator` because they
are likely rendered together (based on my read of the
`Hyrax::PresenterRender`).  Whereas the `ubiquity/creator` partial is
rendered through other pathways.

```shell
> rg "section ==" ~/git/britishlibrary

app/views/records/show_fields/_creator.html.erb:
<button id="collection-details-creator-collapse" class="<%= section == 'header' ? 'btn-secondary' : 'collapse-fields' %>" data-toggle="collapse" data-target=".collection-details-creator.collapse" aria-expanded="false" aria-controls="collapse">

app/views/records/show_fields/_contributor.html.erb:
<button id="collection-details-contributor-collapse" class="<%= section == 'header' ? 'btn-secondary' : 'collapse-fields' %>" data-toggle="collapse" data-target=".collection-details-contributor.collapse" aria-expanded="false" aria-controls="collapse">

app/views/shared/ubiquity/creator/_show_array_hash.html.erb:
<button id="<%= section %>-creator-collapse" class="<%= section == 'header' ? 'btn-secondary collapse-fields' : 'collapse-fields' %>" data-toggle="collapse" data-target=".<%= section %>-creator.collapse" aria-expanded="false" aria-controls="collapse">
```

Related to:

- https://github.com/scientist-softserv/britishlibrary/issues/284
- https://scientist-inc.sentry.io/issues/3941511830

[1]: https://assaydepot.slack.com/archives/C0313NK2LJ0/p1676636500470559?thread_ts=1676568931.180859&cid=C0313NK2LJ0

[2]: https://github.com/samvera/hyrax/blob/64c0bbf0dc0d3e1b49f040b50ea70d177cc9d8f6/app/presenters/hyrax/presenter_renderer.rb#L39-L45

Modified-by: Jeremy Friesen <jeremy.n.friesen@gmail.com>