Ignore `creator` object for FileSet import
In the British Library code-base there are two declarations of
`creator`:

1. For FileSet the creator is an array of literals (e.g. "Jeremy Friesen")
2. For works the creator is an array of complex objects (e.g. [{ given_name:
   "Jeremy", family_name: "Friesen" }])

This creates issues when we use the same parser fields for FileSet and
non-FileSet; namely we attempt to cast the FileSet creator into a
complex object.  Which fails.

With this commit I'm doing two things:

1. Appeasing rubocop
2. Declaring the FileSet's creator as a non-object field

In declaring FileSet's creator as a non-object field, we avoid the
serialization.

As I think more on this, I would ideally like to have different parsers
for FileSets.  However, I think this will be difficult as we use the
parser to sniff out the object's "factory_class".

Related to:

- https://github.com/scientist-softserv/britishlibrary/issues/289
- https://github.com/samvera-labs/bulkrax/issues/754
- https://github.com/samvera-labs/bulkrax/issues/747