Handle creator export disparity for FileSet and works
This commit includes two primary things:

- CsvParser fix to logic (see https://github.com/samvera-labs/bulkrax/pull/756)
- CsvEntry fix for creator property disparity between FileSets and other works

Without this hack, when we export a FileSet, it skips exporting the
FileSet's creator.  This is because we are overloading the `creator` in
the parser.  With this change, the complex `creator` object for works
will export correctly (e.g. have the constituent field parts in the
CSV).  And the FileSet will have a `creator` column.

Related to:

- #289
- https://github.com/samvera-labs/bulkrax/pull/756