Update IIIF Print and Hyrax
Hyrax 2.9.6 has a very restrictive dependency on the IIIF Manifest gem
which does not have presentation version 3 manifest thumbnail support.
This points to a branch of Hyrax that expands the version of of the
manifest gem and makes v1.3.1 available.  Also in this commit, updating
IIIF Print to the latest.