Modify date accepted logic
Added an additional guard so [""] doesn't throw an error.  This occurs
usually with Bulkrax CSV imports when an entry has a header but no value
in the field.  Also the #formatted_date method doesn't exist anywhere so
it was removed.  The only unknown is if "date_published_dtsim" is ever a
case vs "date_published_tesim".  In any case, in the old logic, if we
had "date_published_dtsim" then we would have gotten a NoMethodError.