Patch service to account for PlanSFunder changes
Correct access to PlanSFunder data to use only a subset of the doi.

Based on PR https://github.com/scientist-softserv/britishlibrary/pull/391
which changed the data stored in the PlanSFunder table for doi. (It now
only saves the numeric portion of the doi due to various options for
the doi, and the need to match regardless of the full url.)