🎁 Add PDF filename and page number to metadata
This commit brings in an decorator for `IiifPrint::Metadata` to add the
PDF filename and page number to the canvas metadata.  The path to the
actual overrode file in the gem is `lib/iiif_print/metadata.rb` but the
decorator is in `app/services/iiif_print/metadata_decorator.rb`.  This
is because the load order was not working when I put the decorator in
the lib folder.

- Resolves: https://github.com/scientist-softserv/britishlibrary/issues/431