🎁 Remove `wholeImageLowResAsJpg` download option
This commit is a bit of a workaround.  By setting the
`confinedImageSize` to a very large number, we can effectively hide the
`wholeImageLowResAsJpg` download option because of the way the.  This
doesn't solve the root problem that the `EXTERNAL_IIIF_URL` is not being
passed to the Universal Viewer correctly.  The "Current view" download
option that appear when the image is zoomed in is still broken but it
maybe not be as used or seen by users as the `wholeImageLowResAsJpg`
download option.

- Todo: Figure out how to actually pass the `EXTERNAL_IIIF_URL` to the
  Universal Viewer.
- See: https://github.com/UniversalViewer/universalviewer/blob/v3.1.1/src/extensions/uv-seadragon-extension/DownloadDialogue.ts#L610-L616
- Resolves: https://github.com/scientist-softserv/britishlibrary/issues/429